+++ example/src/sine-wave-frame-buffer-filler.ts
@@ -2,29 +2,27 @@ import type { FrameBufferFiller, FrameBufferWriter } from '@ain1084/audio-workle
 import { SineWaveGenerator } from './sine-wave-generator'
 
 export type FillerParameters = Readonly<{
-  frequency: number
   sampleRate: number
+  frequencies: number[]
 }>
 
 export class SineWaveFrameBufferFiller implements FrameBufferFiller {
-  private readonly _generator: SineWaveGenerator
-  private readonly _tempBuffer = new Float32Array(128)
+  private readonly _generators: SineWaveGenerator[]
   constructor(params: FillerParameters) {
-    this._generator = new SineWaveGenerator({ frequency: params.frequency, sampleRate: params.sampleRate })
+    this._generators = params.frequencies.map((frequency) => {
+      return new SineWaveGenerator({ ...params, frequency })
+    })
   }
 
   fill(writer: FrameBufferWriter): boolean {
-    writer.write((frame) => {
-      let totalFrames = 0
-      while (totalFrames < frame.frames) {
-        const sectionFrames = Math.min(this._tempBuffer.length, frame.frames - totalFrames)
-        for (let i = 0; i < sectionFrames; ++i) {
-          this._tempBuffer[i] = this._generator.nextValue() * 0.3
-        }
-        frame.buffer.setFrames(frame.index + totalFrames, this._tempBuffer, 0, sectionFrames)
-        totalFrames += sectionFrames
+    const samplesPerFrame = this._generators.length
+    writer.write((buffer) => {
+      for (let i = 0; i < buffer.length; i += samplesPerFrame) {
+        this._generators.forEach((generator, channelIndex) => {
+          buffer[i + channelIndex] = generator.nextValue() * 0.3
+        })
       }
-      return frame.frames
+      return buffer.length / samplesPerFrame
     })
     return true
   }