+++ src/write-strategy/worker/worker.ts
@@ -1,4 +1,4 @@
-import { createFrameBufferWriter, FillerFrameBufferConfig } from '../../frame-buffer/buffer-factory'
+import { createFrameBufferWriter, FillerFrameBufferConfig } from '../../frame-buffer/buffer-config'
 import type { FrameBufferFiller } from '../../frame-buffer/buffer-filler'
 import type { FrameBufferWriter } from '../../frame-buffer/buffer-writer'
 import { MessageToStrategy, MessageToWorker } from './message'